Refactor KSZ8851 Ethernet SPI driver
Cleave out the actual SPI and IRQ handling machinery into a `Provider`
class, and provide that class for Sunburst.

This changes the names of some things to better namespace them, but
that's meant to be purely syntactic.
